requirements.txt
click
pycp2110